requirements.txt
Pillow~=10.0.0
numpy~=1.24.3
pywin32~=306
dacite~=1.8.0